Added origin marker, refactored model view control
