+++ game/assets/models/general/origin_marker/idle.smd
@@ -1,0 +1,8 @@
+version 1
+nodes
+  0 "root"  -1
+end
+skeleton
+time 0
+  0 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
+end
