+++ game/assets/models/general/origin_marker/origin_marker.qc
@@ -1,0 +1,14 @@
+$cd .
+$cdtexture "../../../../content/nfopen/textures"
+$modelname "origin_marker.mdl"
+
+$root "root"
+$scale 16
+$noembeddedtextures
+
+$bodygroup "studio"
+{
+	studio "origin_marker"
+}
+
+$sequence "idle" "idle" fps 20 loop
