+++ game/content/nfopen/models/origin_marker.mdl
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c9aa227db78fcd276e479854a3b5c3b1c169ec71992f577a2ebedbc35fc2c541
+size 2864
