+++ game/content/nfopen/textures/mdl/tools/origin_marker.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:78956aa58d2e08f278f335e71152a72238016f905bcf56c4c4420c8352d10b52
+size 7710
